+++ src/main/java/com/myo/madrasati/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.myo.madrasati.controllers;
+package com.myo.madrasati.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
